feat: support writing snapshots

+++ src/Be.Vlaanderen.Basisregisters.AggregateSource/SnapshotContainer.cs
@@ -9,7 +9,7 @@ namespace Be.Vlaanderen.Basisregisters.AggregateSource
 
     public class SnapshotInfo
     {
-        public int Position { get; set; }
+        public long Position { get; set; }
 
         public string Type { get; set; }
     }
